% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation_indices_reference.R
\name{evmk_sums_reference}
\alias{evmk_sums_reference}
\title{Calculate potential evaporation (Makkink), year mean (seasons sum) and summer for brochure validation/ only for reference period}
\usage{
evmk_sums_reference(ofile = "uitvoer.txt")
}
\arguments{
\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}
}
\description{
Function reads Makkink evaporation reference series
and makes season sums
}
