% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation_indices.R
\name{evmk_sums_relchange}
\alias{evmk_sums_relchange}
\title{Calculate potential evaporation (Makkink), year mean (seasons sum) and summer for brochure validation}
\usage{
evmk_sums_relchange(input_tg, input_rsds, scenario, horizon = NA,
  regions = "NLD", ofile = NA)
}
\arguments{
\item{input_tg}{input file for tg}

\item{input_rsds}{input file for rsds}

\item{scenario}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{horizon}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{regions}{vector of regions
 KNMI14 distinguishes following regions:\cr
 <NLD> Nederland (DEFAULT) \cr
 <NWN> Noordwest Nederland \cr
 <ZWN> Zuidwest Nederland \cr
 <NON> Noordoost Nederland \cr
 <MON> Middenoost Nederland \cr
<ZON> Zuidoost Nederland}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to input}
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate, makes season sums and summer values
}
