% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationIndices.R
\name{PrecIndicesWrapper}
\alias{PrecIndicesWrapper}
\title{Calculates a set of precipitation related indices (thresholds) for all scenarios, horizons, and
seasons at once}
\usage{
PrecIndicesWrapper(input, subscenario = "centr")
}
\arguments{
\item{input}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}

\item{subscenario}{subscenario for extreme precipitation
("lower", "centr" (=DEFAULT), "upper")}
}
\description{
Calculates a set of daily precipitation related indices as
  they were defined for the KNMI14 scenarios brochure
}
