% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_highest_precipitation.R
\name{PrecipDeficit}
\alias{PrecipDeficit}
\title{Calculate the mean highest precipitation deficit during Growing Season (GS) (April-August)
as it was calculated for KNMI14 scenarios brochure
this index uses the evmk of DeBilt and precipitation amount of 102 stations (the homogenenised ones) over the NL}
\usage{
PrecipDeficit(inputTemp, inputRad, inputPrec, scenario, horizon = 2030,
  regions = "NLD")
}
\arguments{
\item{inputTemp}{input file for tg}

\item{inputRad}{input file for rsds}

\item{inputPrec}{input for rr}

\item{scenario}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{horizon}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{regions}{vector of regions
KNMI14 distinguishes following regions:\cr
<NLD> Nederland (DEFAULT) \cr
<NWN> Noordwest Nederland \cr
<ZWN> Zuidwest Nederland \cr
<NON> Noordoost Nederland \cr
<MON> Middenoost Nederland \cr
<ZON> Zuidoost Nederland}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to input}
}
\description{
function reads reference ts for evmk, rr, tg & rdrs
}
