% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_highest_precipitation_reference.R
\name{PrecipDeficit_ref}
\alias{PrecipDeficit_ref}
\title{Calculate the mean highest precipitation deficit during Growing Season (GS) (April-August)
as it was calculated for KNMI14 scenarios brochure For Just the REFERENCE Period
this index uses the evmk of DeBilt and precipitation amount of 102 stations (the homogenenised ones) over the NL}
\usage{
PrecipDeficit_ref(ofile = "uitvoer.txt")
}
\arguments{
\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}
}
\description{
function reads reference ts for evmk, rr, tg & rdrs
}
