% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_highest_precipitation.R
\name{PrecipDeficit_sce}
\alias{PrecipDeficit_sce}
\title{Calculate the mean highest precipitation deficit during Growing Season (GS) (April-August)
as it was calculated for KNMI14 scenarios brochure
this index uses the evmk of DeBilt and precipitation amount of 102 stations (the homogenenised ones) over the NL}
\usage{
PrecipDeficit_sce(input_tg, input_rsds, input_rr, scenario, horizon = NA,
  ofile = NA, regio.file = NA)
}
\arguments{
\item{input_tg}{input file for tg}

\item{input_rsds}{input file for rsds}

\item{input_rr}{input for rr}

\item{scenario}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{horizon}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to input}

\item{regio.file}{this (optional) argument provides the name of an ASCII file that relates the stations to
a particular region. First column is station id and second column region
               KNMI14 distinguishes following regions:
               <NLD> Nederland [DEFAULT]
               <NWN> Noordwest Nederland
               <ZWN> Zuidwest Nederland
               <NON> Noordoost Nederland
               <MON> Middenoost Nederland
               <ZON> Zuidoost Nederland}
}
\description{
function reads reference ts for evmk, rr, tg & rdrs
}
