% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationIndices.R
\name{PrecipThreshIndices}
\alias{PrecipThreshIndices}
\title{Calculate precipitation threshold indices}
\usage{
PrecipThreshIndices(input, index, season)
}
\arguments{
\item{input}{input file with precipitation}

\item{index}{vector threshold indices}

\item{season}{vector year,winter, spring, summer, autumn}
}
\description{
Calculate precipitation threshold indices
}
