% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgIndices.R
\name{TempAvgIndicesWrapper}
\alias{TempAvgIndicesWrapper}
\title{Calculates a set of TG related indices for all scenarios, horizons, and
seasons at once}
\usage{
TempAvgIndicesWrapper(input, regions = "NLD", ofile = NA)
}
\arguments{
\item{input}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}

\item{regions}{vector of regions
KNMI14 distinguishes following regions:\cr
<NLD> Nederland (DEFAULT) \cr
<NWN> Noordwest Nederland \cr
<ZWN> Zuidwest Nederland \cr
<NON> Noordoost Nederland \cr
<MON> Middenoost Nederland \cr
         <ZON> Zuidoost Nederland}

\item{ofile}{(DEFAULT=NA) Name of the output file to write the indices to.
Format is similar to input without the 5 first lines}
}
\description{
Calculates a set of daily average temperature related indices as
  they were defined for the KNMI14 scenarios brochure
}
