% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txIndices.R
\name{TempMaxIndices}
\alias{TempMaxIndices}
\title{Calculates a set of daily maximum temperature related indices as they were defined
for the KNMI14 scenarios brochure}
\usage{
TempMaxIndices(ifile_tx, index, ofile = NA, scenario, horizon = 2030,
  season, regio.file = NA)
}
\arguments{
\item{index}{indices ("nID", "nWD", "nSD", "nTD", "aTX")}

\item{ofile}{(DEFAULT=NA) Name of the output file to write the indices to.
Format is similar to ifile without the 5 first lines}

\item{scenario}{scenario ("GL", "GH", "WL", "WH")}

\item{horizon}{time horizon ( DEFAULT=2030, 2050, 2085)}

\item{season}{season (0= year, 1=winter, 2=spring, 3=summer, 4=autumn)}

\item{regio.file}{this (optional) argument provides the name of an ASCII file that relates the stations to
a particular region. First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            [DEFAULT]
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}

\item{ifile}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}
}
\description{
function calculates a set of TX related incides
}
