% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation_indices.R
\name{evmk_sums_relchange}
\alias{evmk_sums_relchange}
\title{Calculate potential evaporation (Makkink), year mean (seasons sum) and summer for brochure validation}
\usage{
evmk_sums_relchange(ifile_ref, ifile_scenario, ofile = "uitvoer.txt", sc,
  p = NA)
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate, makes season sums and summer values
}

