% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation_indices.R
\name{evmk_sums_relchange}
\alias{evmk_sums_relchange}
\title{Calculate potential evaporation (Makkink), year mean (seasons sum) and summer for brochure validation}
\usage{
evmk_sums_relchange(ifile_tg, ifile_rsds, ofile = "uitvoer.txt", sc, p = NA,
  regio.file = NA)
}
\arguments{
\item{ifile_tg}{input file for tg}

\item{ifile_rsds}{input file for rsds}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}

\item{sc}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{p}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{regio.file}{this (optional) argument provides the name of an ASCII file that relates the stations to
a particular region. First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            [DEFAULT]
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate, makes season sums and summer values
}

