% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationIndices.R
\name{PrecIndicesWrapper}
\alias{PrecIndicesWrapper}
\title{Wrapper function for precipitation indices}
\usage{
PrecIndicesWrapper(input)
}
\arguments{
\item{input}{precipitation input file}
}
\description{
Wrapper function for precipitation indices
}
