% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationIndices.R
\name{PrecipThreshIndices}
\alias{PrecipThreshIndices}
\title{precipitation threshold index}
\usage{
PrecipThreshIndices(input, threshold, scenario = NA, horizon = NA, season,
  subscenario, ofile = NA)
}
\arguments{
\item{index}{indices ("nID", "nWD", "nSD", "nTD", "aTX")}
}
\description{
Calculates a set of daily precipitation related indices as
they were defined for the KNMI14 scenarios brochure
}
